+++ ModulePhysics.h
@@ -47,7 +47,7 @@ public:
 	update_status PostUpdate();
 	bool CleanUp();
 
-	PhysBody* CreateCircle(int x, int y, int radius, int type, int restitution = 0, SDL_Texture* text = nullptr, bodyType myType=NONE);
+	PhysBody* CreateCircle(int x, int y, int radius, int type, float restitution = 0, SDL_Texture* text = nullptr, bodyType myType=NONE);
 	PhysBody* CreateCircleSensor(int x, int y, int radius, bodyType myType = NONE);
 	PhysBody* CreateRectangle(int x, int y, int width, int height, int type, SDL_Texture* text = nullptr, bodyType myType = NONE);
 	PhysBody* CreateRectangleSensor(int x, int y, int width, int height, bodyType myType = NONE);
